Refector userlist code
